Move span to utils directory.
