+++ ngraph/core/reference/include/ngraph/runtime/reference/utils/span.hpp
